\name{test}
\alias{test}
\docType{data}
\title{
  A simulated testing dataset.
}
\description{
The dataset contains the information on disease status and values from three biomarkers.
This dataset is used as testing sample.
}
\usage{
data(test)
}
\format{
A data frame containing the following variables for each of 300 observations: status (disease status coded as 1 for case or 0 for control), values from three biomarkers.
}
\examples{
  data(test);
}
\keyword{datasets}
